Js to Tsx conversions and search.js modification with fetch.
