+++ components/editProgram/Mastered.tsx
@@ -1,0 +1,71 @@
+import React, { FC, useEffect, useState } from 'react'
+import Behavior from "./Behavior";
+
+interface BehaviorItem
+ { 
+  
+  id: number; 
+  name: string; 
+  description: string; 
+  type: string; 
+  domain: string []; 
+  masteryCriteria: string; 
+  mastered: boolean; 
+
+}
+
+interface MasteredProps 
+{
+  studentID: string; 
+}
+
+
+
+
+
+
+const Mastered = ({ studentID }) => {
+    
+    const [behavior, setBehavior] = React.useState([]);
+  
+    React.useEffect(() => {
+      setBehavior([
+        {
+          id: 0,
+          name: "Touch Toes",
+          description: "Test if client is able to touch toes",
+          type: "Trial",
+          domain: ["Behaviors for Increase"],
+          masteryCriteria: "5 consecutive days passed.",
+          mastered: false,
+        },
+        {
+          id: 1,
+          name: "Punching",
+          description:
+            "If student punches anyone",
+          type: "frequency",
+          domain: ["Behaviors for decrease"],
+          masteryCriteria: "2 consecutive days passed.",
+          mastered: false,
+        },
+        {
+          id: 2,
+          name: "Responding to Name",
+          description: "Reacts when name is called",
+          type: "probe",
+          domain: ["Listener Reponding"],
+          masteryCriteria: "10 Sessions Passed",
+          mastered: true,
+        },
+      ]);
+    }, []);
+    return (
+        <div>
+            <Behavior list={behavior} />
+        </div>
+    )
+}
+
+export default Mastered
+
